third_party/blink/web_tests/fast/css/sibling-selectors-expected.txt

test0 foo

test0 bar

test1 foo

test1 bar

test2 foo

test2 bar



#test0 :first-of-type
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


#test0 :last-of-type
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(255, 0, 0)'


#test1 :first-of-type
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


#test1 :last-of-type
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(255, 0, 0)'


#test2 :first-of-type
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'


#test2 :last-of-type
PASS document.defaultView.getComputedStyle(element, null).getPropertyValue('background-color') is 'rgb(255, 0, 0)'
PASS successfullyParsed is true

TEST COMPLETE
